+++ src/Domain/Entities/Transactions/Transaction.cs
@@ -6,14 +6,14 @@ public class Transaction
 {
     public int Id { get; set; }    
     public int DateCreated { get; set; }
-    public int MinSpentAmount { get; set; }
+    public decimal MinSpentAmount { get; set; }
     public int MinSpentAmountPoints { get; set; }
-    public int UpperRangeSpentAmount { get; set; }
+    public decimal UpperRangeSpentAmount { get; set; }
     public int UpperRangeSpentPoints { get; set; }
-    public int SubTotal { get; set; }
+    public decimal SubTotal { get; set; }
     public int TotalVAT { get; set; }
-    public int GrandTotal { get; set; }
-    public int TaxRate { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal TaxRate { get; set; }
     public int CustomerId { get; set; }
     public virtual Customer Customer { get; set; }
